fix type mismatch

+++ src/tests/shamrock/tree/key_pair_sortTests.cpp
@@ -100,7 +100,7 @@ void wrapper_bench_key_sort(std::string name) {
 
     std::vector<f64> test_sz;
     for (u32 i = 16; i < lim_bench; i *= 2) {
-        test_sz.push_back(i);
+        test_sz.push_back(f64(i));
     }
 
     auto &res = shamtest::test_data().new_dataset(name);
